Fixes MmTelFeatureConnection regression during creation
The MmTelFeatureConnection creation process involves notifying the
MmTelFeature if the connected MmTelFeature is available.
The notification of this had been moved to an executor, which
caused the update to happen asynchronously after the creation. Since
the default state of the MmTelFeatureConnection is "not available",
this was causing the ImsManager.Connector to retry multiple times.

Set the state of mIsAvailable to true when created, since we will only
create the MmTelFeatureConnection if the Binder is already available.

Bug: 139500829
Bug: 139171862
Test: atest GtsTelephonyTestCases
Change-Id: Ic2be4984fc62fc85da641cebe36533d3fca6717b
Merged-In: Ic2be4984fc62fc85da641cebe36533d3fca6717b