Standardize RTT behavior across carriers
Remove references to the ignore rtt setting carrier config key, and
treat it as always-on for all carriers. Also turn on RTT in the IMS
stack when there's no SIM card inserted so that emergency calls can use
RTT in that scenario.

Bug: 122250491
Bug: 130010750
Test: manual
Change-Id: Ibf9cf45041f819a28757456082a1a80176797e81
Merged-In: Ibf9cf45041f819a28757456082a1a80176797e81