Create method to query capability status directly on ImsServices
There was a race case that occurs when an ImsService changes its capability status
and with a call against IsAvailable(capability) within the telephony component which
checks a cache.  This change moves the call in ImsPhoneCallTracker from its internal
cache to instead check the capability status directly within the ImsService.

Bug: 149780842
Test: android.telephony.ims.cts.ImsServiceTest
Change-Id: Ia5329bdb0c2338c9d9b215d74ad1a027fb7f5ba6