Move to a push model of querying ImsFeature Binders
[RESUBMISSION AFTER REVERT aosp/1426022]

Instead of a polling model, allow listeners to instead
listen for updates to a specific ImsFeature interface
and get notified when they are available.

FeatureConnectionRepository is used by the ImsServiceControllers
to store new MMTEL and RCS feature binders along with the
associated IMS config and registration interfaces that should be
used along with them. Listeners can then register for updates to
the ImsFeatures, including changes to capabilities and state.

The FeatureConnector has also been refactored to listen ann
update MmTel and RCS FeatureConnections in the ImsManager and
RcsFeatureManager instances as Binders become
available/unavailable or change when new SIM cards are
inserted/removed instead of using a polling method of
retrieving Binder instances.

Fixes: 169864814
Test: atest ImsCommonTests
Change-Id: Ie35b443ee2734c72f52d736f3980c1521a958d96