Increase ImsManager efficiency when using getInstance
1) Remove FeatureConnector "oneshot" idea, instead always
use the normal #connect method.

2) FeatureConnector should not call back while holding lock.

3) ImsManager#getInstance has been reworked to increase efficiency.
Instead of creating a new ImsManager every time getInstance is called,
create one ImsManager per slot and share it amongst all users. This
is due to ImsManager being a pretty "heavyweight" object (may spawn
threads). In order to accomplish this with the new FeatureConnector,
each ImsManager will be contained in a static InstanceManager class,
which will manage internally updating the MmTelFeatureConnection for
the ImsManager instance as it is updated.

Test: atest ImsCommonTests
Merged-In: Idc56628ea7b6cf705ce55d163ea29e5807a2515c
Change-Id: Idc56628ea7b6cf705ce55d163ea29e5807a2515c