Fix MmTelFeatureConnection closing causing bad state
MmTelFeatureConnection should only tear down listeners associated
with #open when #close is called and not also tear down unrelated
state.

This refactors the #open method to initialize mmtel/ecbm/multiendpoint
listeners so that they can be torn down properly when #close is called.

The UT interface is not tied to open/close, but it is also limited to
a single listener. Therefore, it is only torn down when the
MmTelFeatureConnection is removed.

Bug: 175810194
Test: atest CtsTelephonyTestCases ImsCommonTests
Change-Id: If42e113ba95cc5157e0fc1688d47edf4cb81831f