Add the ability to include RCS capabilities in SIP PUBLISH
1) Introduce the PublishServiceDescTracker, which manages the
mapping between the service-description elements in the PIDF XML and
the IMS feature tags that are associated with them. Based on the
IMS registration received from the ImsService, it adjusts the
RCS capabilities that should be reported to the network as well as
allowing customizable mappings on a per carrier basis.

2) Integrate PublishServiceDescTracker into DeviceCapabilityInfo,
which uses the latest IMS registration information to generate the
capabilities structure used by PIDF parsing code to generate the
PIDF XML and ultimately send it to the network in a SIP PUBLISH message.

3) Introduce better logging into SIP publish to verify this
information.

Bug: 180868933
Test: atest ImsCommonTests
Change-Id: Ie9a34715dbe8577471fd39c1a01d19b6b000bdf9