[UCE] Reset the retry count and times related publish when the publish status changed to not_published
we do not reset the PUBLISH throttling timer when CapabilityExchangeEventListener#onUnpublish is called and PDN is torn down. This means that PUBLISH is delayed up to a minute when IMS PDN comes back up. Instead, we should be resetting this timer when the onUnpublish event is received

Bug: http://b/204198074
Test: atest PublishControllerImplTest
Change-Id: I3a01bf33921922fcc440fc5d8f4efb3d99916b28
Merged-In: I3a01bf33921922fcc440fc5d8f4efb3d99916b28