Fixing empty tuple capability always expired
EAB stores capabilities in two different tables (common, presence), and
stores the expiration time in the Presence table, because a contact may
have multiple tuples, and each tuple may have its own expiration time.

But if the server sends capabilities with empty tuples, EAB will not
store any data in the Presence table. It causes the EAB to treat it as
an expired capability and request a new capability every time.

[Fix] If there is no tuple ability, store the current timestamp and null
presence attributes in the presence table.

Bug: 203676348
Test: atest EabControllerTest
Change-Id: Id54a71b71d353f3d2fa5c3246ab534ad581fe245