Shortcut and return caps of contacts who have non-expired cap info
When an application requests the UCE capabilities of a list of
numbers, return the cached info from the EAB DB right for numbers
where the cap info hasnt expired yet instead of queueing the request
for those numbers behind other pending  network requests. This
helps improve latency issues that were happening when network
requests were taking a long time (1 min+) to finish on the network.

Bug: 205878771
Test: atest ImsCommonTests CtsTelephonyTestCases:RcsUceAdapterTest
Change-Id: Id7f07cbf3107271c7b2e9b9396d2c795974f3311