Fix RTT NPE due to ImsCall#close being called
When the IMS call is closed, in some cases, a pending RTT
initialization request could still be in the queue to be
processed. Since this depends on resources that have since been
release, it can cause an NPE.

Guard against this condition and return early if this condition
is hit.

Bug: 236279778
Test: atest FrameworksTelephonyTests:ImsCallTest
Change-Id: I1c734eebd8135655752e646c5967e65cfe140c89