Remove airplane mode handling code.
In VZW TC 3.18, the device did not need to send a PUBLISH when a 412 response was received for a UN PUBLISH request.
Turn on airplane mode for testing. As a result, the device sends a UN PUBLISH request. However, since airplane mode has changed, the Framework sends a PUBLISH request to the ImsService.
So this case failed.

Bug: b/243137045
Test: Turn on airplane mode > deice sends a UN PUBLISH request > Not handle airplane mode change event in the publish controller.
Test: atest DeviceCapabilityListenerTest

Change-Id: Id96bd4a233e9208865afaa9c3fe0cccb239bad93