If only RCS or MMTEL is unregistered, send a modify PUBLISH request
When RCS or MMTEL is unregistered, send a modify PUBLISH request message even if the associated URI has not changed.
When RCS or MMTEL change from Unregistered to Unregistered, the registration status does not change.
In this case, the modify PUBLISH request is sent depending on whether the associated URI has been changed.

Bug: b/235973951
Test: atest DeviceCapabilityListenerTest
Test: TMO Live network. The RCS and MMTEL are registered but the associated URI of RCS is not set. After that, RCS will be unregistered. Verified that the modify PUBLISH request has been sent to the netowrk.
Change-Id: I96b565c15f11ee1083b56db29971ddee930d7e64