Add exported flag to exposed runtime receiver
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U
and registering a receiver for non-system broadcasts must
specify either the exported or not exported flag when invoking
registerReceiver; if one of these flags is not provided, the
platform will throw a SecurityException. This commit updates
the receiver in DeviceCapabilityListener with the RECEIVER_EXPORTED
flag since it is registering for the TTY_PREFERRED_MODE_CHANGED
action which can be sent from UIDs other than the system; the
receiver is also protected by a signature permission which limits
the apps that could send this broadcast to this receiver.

Fixes: 266769311
Test: atest CellBroadcastUiTest#testAlertUiOnReceivedAlert
Change-Id: I821e92fdd0a3027b189a6cdec5da7740ea8846e1